orchestrator/orchestrator.py
"""
Main orchestrator class that manages the neural architecture search lifecycle.

The Orchestrator is the central control unit of the neural architecture search process.
It coordinates between different components:
- Requests new architectures from search strategies
- Triggers code generation via LLM
- Schedules training runs in containers
- Records and analyzes results

Example:
    >>> config = Config()
    >>> orchestrator = Orchestrator(config)
    >>> orchestrator.run_cycle()  # Runs one complete search cycle
"""

from typing import Dict, Any, Optional, List
from pathlib import Path
import asyncio
import logging

from ..arch_space import ArchitectureVector
from ..llm_code_gen import CodeGenerator
from ..env_manager import ContainerManager
from ..results_db import ResultsDB
from ..utils.logging import setup_logger
from ..utils.version_control import VersionControl

logger = setup_logger(__name__)


class Orchestrator:
    """
    Manages the lifecycle of experiments: requests new architectures,
    triggers code generation, schedules training runs, and records results.

    The orchestrator implements the main control loop of the architecture search:
    1. Get next architecture from search strategy
    2. Generate code via LLM
    3. Run experiment in container
    4. Record results and update search strategy

    Attributes:
        config (Dict[str, Any]): Configuration dictionary containing:
            - search_strategy: Strategy class name or instance
            - llm_provider: LLM provider configuration
            - container_config: Container runtime settings
            - experiment_defaults: Default training parameters
    """

    def __init__(self, config: Dict[str, Any]):
        """
        Initialize the orchestrator with configuration.

        Args:
            config: Configuration dictionary with all necessary settings
        """
        self.config = config
        self._setup_components()

    def _setup_components(self) -> None:
        """Initialize all required components from configuration."""
        raise NotImplementedError

    async def run_cycle(self) -> Dict[str, Any]:
        """
        Run a single cycle of the neural architecture search.

        A cycle consists of:
        1. Getting next architecture to evaluate
        2. Generating implementation code
        3. Running the experiment
        4. Recording results

        Returns:
            Dict containing cycle results including:
            - architecture_id: Unique identifier for the architecture
            - metrics: Performance metrics from training
            - code_version: Git commit hash of generated code

        Raises:
            RuntimeError: If any component fails critically
        """
        raise NotImplementedError

    async def get_next_architecture(self) -> Dict[str, Any]:
        """
        Get the next architecture to evaluate from the search strategy.

        Returns:
            Dict containing:
            - vector: The architecture vector
            - metadata: Any strategy-specific metadata

        Raises:
            ValueError: If strategy returns invalid architecture
        """
        raise NotImplementedError

    async def submit_results(self, results: Dict[str, Any]) -> None:
        """
        Submit results from a completed experiment.

        Args:
            results: Dictionary containing:
                - architecture_id: ID of the evaluated architecture
                - metrics: Dictionary of performance metrics
                - training_time: Time taken for training
                - resource_usage: CPU/GPU/memory usage stats
        """
        raise NotImplementedError

    async def run_batch(
        self, batch_size: int, parallel: bool = True
    ) -> List[Dict[str, Any]]:
        """
        Run multiple architecture evaluations in parallel.

        Args:
            batch_size: Number of architectures to evaluate
            parallel: If True, run evaluations concurrently

        Returns:
            List of results from each evaluation
        """
        raise NotImplementedError
